Refactor tests: Replace UniversalParser with Parser and remove obsolete test_base_parser
- Deleted test_base_parser.py as it was no longer needed.
- Created test_parser.py to thoroughly test the new Parser class functionality.
- Updated test_relative_links.py to use the Parser class instead of UniversalParser.
- Modified test_cli.py to reflect changes in the CLI commands, replacing UniversalParser with Parser.
