+++ tapio/parsers/parser.py
@@ -1,11 +1,13 @@
-"""Universal HTML content parser module.
+"""HTML content parser module.
 
-This module contains the UniversalParser class, which loads site-specific
+This module contains the Parser class, which loads site-specific
 configurations and extracts content from HTML pages accordingly.
 """
 
+import json
 import logging
 import os
+from datetime import datetime
 from pathlib import Path
 from typing import Any
 from urllib.parse import urljoin
@@ -15,16 +17,15 @@ import yaml
 from lxml import html
 
 from tapio.config.settings import DEFAULT_DIRS
-from tapio.parsers.base_parser import BaseParser
 from tapio.parsers.config_models import (
     ParserConfigRegistry,
     SiteParserConfig,
 )
 
 
-class UniversalParser(BaseParser):
+class Parser:
     """
-    Universal HTML content parser that uses site-specific configurations.
+    HTML content parser that uses site-specific configurations.
 
     This parser loads configurations for different websites and extracts
     content using the appropriate selectors for each site.
@@ -38,7 +39,7 @@ class UniversalParser(BaseParser):
         config_path: str | None = None,
     ):
         """
-        Initialize the universal parser.
+        Initialize the parser.
 
         Args:
             site: Site to parse (must match a key in config)
@@ -49,14 +50,87 @@ class UniversalParser(BaseParser):
         self.site = site
         self.config = self._load_site_config(site, config_path)
         self.current_base_url: str | None = None  # Will store the base URL of the current document
+        
+        self.input_dir = input_dir
+        self.output_dir = os.path.join(output_dir, self.config.site_name or "default")
+        self.site_name = self.config.site_name
+        self.setup_logging()
 
-        super().__init__(
-            input_dir=input_dir,
-            output_dir=output_dir,
-            site_name=self.config.site_name,
+        # Load URL mappings if available
+        self.url_mappings: dict[str, dict[str, str]] = {}
+        self._load_url_mappings()
+
+        # Create output directory if it doesn't exist
+        os.makedirs(self.output_dir, exist_ok=True)
+
+        logging.info(f"Initialized Parser for {self.config.site_name}")
+
+    def setup_logging(self) -> None:
+        """Set up logging configuration"""
+        logging.basicConfig(
+            level=logging.INFO,
+            format="%(asctime)s [%(levelname)s] %(message)s",
+            handlers=[logging.StreamHandler()],
         )
+        self.logger = logging.getLogger(__name__)
 
-        logging.info(f"Initialized UniversalParser for {self.config.site_name}")
+    def _load_url_mappings(self) -> None:
+        """Load URL mappings from the JSON file"""
+        mapping_file = os.path.join(self.input_dir, "url_mappings.json")
+        if os.path.exists(mapping_file):
+            try:
+                with open(mapping_file, encoding="utf-8") as f:
+                    self.url_mappings = json.load(f)
+                self.logger.info(f"Loaded {len(self.url_mappings)} URL mappings")
+            except Exception as e:
+                self.logger.error(f"Error loading URL mappings: {str(e)}")
+        else:
+            self.logger.warning(f"URL mapping file not found: {mapping_file}")
+            # Still continue processing - URL mappings are optional
+
+    def _get_original_url(self, file_path: str | Path) -> str | None:
+        """
+        Get the original URL for a file path from the URL mappings.
+
+        Args:
+            file_path: Path to the HTML file
+
+        Returns:
+            Original URL or None if not found
+        """
+        # Handle Path objects by converting to string
+        file_path_str = str(file_path)
+
+        # Try to find the file in our mappings
+        for key, value in self.url_mappings.items():
+            # Check if the path ends with our key
+            if file_path_str.endswith(key):
+                return value.get("url")
+
+        # Try to infer the key from the file path
+        try:
+            # Extract relative path from input_dir
+            rel_path = os.path.relpath(file_path, self.input_dir)
+
+            # Check for both slash directions (OS compatibility)
+            if rel_path in self.url_mappings:
+                return self.url_mappings[rel_path].get("url")
+
+            # Try with forward slashes (URL style)
+            rel_path_fwd = rel_path.replace("\\", "/")
+            if rel_path_fwd in self.url_mappings:
+                return self.url_mappings[rel_path_fwd].get("url")
+
+            # Try with the filename only
+            filename = os.path.basename(file_path)
+            for key, value in self.url_mappings.items():
+                if key.endswith(filename):
+                    return value.get("url")
+        except Exception as e:
+            self.logger.debug(f"Error finding URL mapping for {file_path}: {e}")
+
+        self.logger.debug(f"No URL mapping found for {file_path}")
+        return None
 
     def _load_site_config(self, site: str, config_path: str | None = None) -> SiteParserConfig:
         """
@@ -274,39 +348,6 @@ class UniversalParser(BaseParser):
 
         return markdown_text
 
-    @classmethod
-    def list_available_site_configs(cls, config_path: str | None = None) -> list[str]:
-        """
-        List all available site configurations.
-
-        Args:
-            config_path: Optional path to custom config file
-
-        Returns:
-            List of available site configuration keys
-        """
-        config_registry = cls._load_config_registry(config_path)
-        return list(config_registry.sites.keys())
-
-    @classmethod
-    def get_site_config(
-        cls,
-        site: str,
-        config_path: str | None = None,
-    ) -> SiteParserConfig | None:
-        """
-        Get detailed information about a specific site configuration.
-
-        Args:
-            site: Site to get configuration for
-            config_path: Optional path to custom config file
-
-        Returns:
-            SiteParserConfig for the specified site, or None if not found
-        """
-        config_registry = cls._load_config_registry(config_path)
-        return config_registry.sites.get(site)
-
     def parse_file(self, html_file: str | Path) -> dict[str, Any] | None:
         """
         Parse a single HTML file from the configured domain.
@@ -346,8 +387,157 @@ class UniversalParser(BaseParser):
                 self.current_base_url = self.config.base_url
                 self.logger.info(f"Using base URL fallback: {self.current_base_url}")
 
-        # Call the parent method to continue with parsing
-        return super().parse_file(html_file)
+        # Parse the file
+        html_file_path = Path(html_file)
+        self.logger.info(f"Parsing {html_file_path}")
+
+        try:
+            # Read the HTML content
+            with open(html_file_path, encoding="utf-8") as f:
+                html_content = f.read()
+
+            # Get relative path from input directory for generating output file path
+            try:
+                rel_path = html_file_path.relative_to(self.input_dir)
+                # Extract the domain from the relative path (first directory)
+                domain_parts = rel_path.parts
+                if len(domain_parts) > 0:
+                    domain = domain_parts[0]
+                else:
+                    domain = "unknown"
+            except ValueError:
+                # If file is not inside input_dir, use the filename
+                rel_path = Path(html_file_path.name)
+                domain = "unknown"
+
+            # Generate a filename for the output markdown
+            output_filename = self._get_output_filename(html_file_path)
+
+            # Parse the HTML content
+            title, content = self._parse_html(html_content)
+
+            # Create metadata for the markdown file
+            metadata = self._create_metadata(html_file_path, title)
+
+            # Save the content as Markdown with frontmatter
+            output_path = self._save_markdown(output_filename, title, content, metadata)
+
+            return {
+                "source_file": str(html_file_path),
+                "output_file": output_path,
+                "title": title,
+                "domain": domain,
+            }
+
+        except Exception as e:
+            self.logger.error(f"Error parsing {html_file_path}: {str(e)}")
+            return None
+
+    def _create_metadata(self, file_path: str | Path, title: str) -> dict[str, Any]:
+        """
+        Create metadata for the markdown file including the original URL.
+
+        Args:
+            file_path: Path to the HTML file
+            title: Title of the page
+
+        Returns:
+            Dictionary with metadata
+        """
+        # Try to get the domain from the path
+        try:
+            rel_path = Path(file_path).relative_to(self.input_dir)
+            domain_parts = rel_path.parts
+            if len(domain_parts) > 0:
+                domain = domain_parts[0]
+            else:
+                domain = "unknown"
+        except ValueError:
+            domain = "unknown"
+
+        # Basic metadata
+        metadata = {
+            "source_file": str(file_path),
+            "title": title,
+            "domain": domain,
+            "parse_timestamp": datetime.now().isoformat(),
+            "parser": self.__class__.__name__,
+        }
+
+        # Add the original URL to the metadata if available
+        original_url = self._get_original_url(file_path)
+        if original_url:
+            metadata["source_url"] = original_url
+
+        return metadata
+
+    def _get_output_filename(self, html_file_path: Path) -> str:
+        """
+        Generate an output filename for the parsed markdown file.
+        Preserves the directory structure from the input directory.
+
+        Args:
+            html_file_path: Path to the source HTML file
+
+        Returns:
+            Output filename with path (without extension)
+        """
+        # Try to extract relative path from the file path
+        try:
+            rel_path = html_file_path.relative_to(self.input_dir)
+            parts = rel_path.parts
+
+            # If there are parts (domain and path), preserve the structure
+            if len(parts) >= 2:
+                # Preserve the original directory structure but replace .html extension
+                rel_output_path = str(rel_path).replace(".html", "")
+                return rel_output_path
+            else:
+                # Just use filename if there's only one part
+                return html_file_path.stem
+        except ValueError:
+            # If the file is not in the input directory, just use its name
+            return html_file_path.stem
+
+    def _save_markdown(
+        self,
+        filename: str,
+        title: str,
+        content: str,
+        metadata: dict[str, Any],
+    ) -> str:
+        """
+        Save content as a markdown file with YAML frontmatter.
+
+        Args:
+            filename: Base filename (without extension)
+            title: Content title
+            content: Markdown content
+            metadata: Dictionary of metadata
+
+        Returns:
+            Path to the saved file
+        """
+        # Ensure filename has .md extension
+        if not filename.endswith(".md"):
+            filename = f"{filename}.md"
+
+        # Create the full output path
+        output_path = os.path.join(self.output_dir, filename)
+
+        # Create parent directories if needed
+        os.makedirs(os.path.dirname(output_path), exist_ok=True)
+
+        # Prepare the markdown content with frontmatter
+        frontmatter = yaml.dump(metadata, default_flow_style=False)
+        markdown_content = f"---\n{frontmatter}---\n\n# {title}\n\n{content}\n"
+
+        # Save the file
+        with open(output_path, "w", encoding="utf-8") as f:
+            f.write(markdown_content)
+
+        self.logger.info(f"Saved markdown to {output_path}")
+        return output_path
 
     def _is_file_in_domain_dir(self, file_path: str | Path) -> bool:
         """
@@ -390,7 +580,97 @@ class UniversalParser(BaseParser):
         self.input_dir = os.path.join(original_input_dir, self.config.base_dir)
 
         try:
-            return super().parse_all()
+            results: list[dict[str, Any]] = []
+
+            # Get all HTML files
+            html_dir = self.input_dir
+
+            # Find all HTML files recursively
+            html_files = []
+            for root, _, files in os.walk(html_dir):
+                for file in files:
+                    if file.endswith(".html"):
+                        html_files.append(os.path.join(root, file))
+
+            self.logger.info(f"Found {len(html_files)} HTML files to parse")
+
+            # Parse each file
+            for html_file in html_files:
+                try:
+                    result = self.parse_file(html_file)
+                    if result:
+                        results.append(result)
+                except Exception as e:
+                    self.logger.error(f"Error parsing {html_file}: {str(e)}")
+
+            # Create an index file for all parsed files
+            if results:
+                self._create_index(results)
+
+            self.logger.info(f"Parsed {len(results)} files")
+            return results
         finally:
             # Restore the original input directory
             self.input_dir = original_input_dir
+
+    def _create_index(self, results: list[dict[str, Any]]) -> str:
+        """
+        Create an index markdown file for all parsed content.
+
+        Args:
+            results: List of parsing results
+
+        Returns:
+            Path to the index file
+        """
+        index_path = os.path.join(self.output_dir, "index.md")
+
+        with open(index_path, "w", encoding="utf-8") as f:
+            f.write(f"# {self.site_name or 'Site'} Parsed Content Index\n\n")
+            f.write(f"Total pages parsed: {len(results)}\n\n")
+            f.write("| Title | Source File | Output File |\n")
+            f.write("|-------|-------------|-------------|\n")
+
+            for result in results:
+                title = result.get("title", "Untitled")
+                source = os.path.basename(result.get("source_file", ""))
+                output = os.path.basename(result.get("output_file", ""))
+
+                # Create relative links to the files
+                f.write(f"| {title} | {source} | [{output}]({output}) |\n")
+
+        self.logger.info(f"Created index at {index_path}")
+        return index_path
+
+    @classmethod
+    def list_available_site_configs(cls, config_path: str | None = None) -> list[str]:
+        """
+        List all available site configurations.
+
+        Args:
+            config_path: Optional path to custom config file
+
+        Returns:
+            List of available site configuration keys
+        """
+        config_registry = cls._load_config_registry(config_path)
+        return list(config_registry.sites.keys())
+
+    @classmethod
+    def get_site_config(
+        cls,
+        site: str,
+        config_path: str | None = None,
+    ) -> SiteParserConfig | None:
+        """
+        Get detailed information about a specific site configuration.
+
+        Args:
+            site: Site to get configuration for
+            config_path: Optional path to custom config file
+
+        Returns:
+            SiteParserConfig for the specified site, or None if not found
+        """
+        config_registry = cls._load_config_registry(config_path)
+        return config_registry.sites.get(site)
